OSX fixes
git-svn-id: https://svn.code.sf.net/p/bzflag/code/branches/v1_7/bzflag@1343 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/bzflag/menus.cxx
@@ -2513,7 +2513,7 @@ void			ServerMenu::setSelected(int index)
       std::vector<std::string> args;
       sprintf(msg, "%d", newPage + 1);
       args.push_back(msg);
-      sprintf(msg, "%d", (servers.size() + NumItems - 1) / NumItems);
+      sprintf(msg, "%ld", (long int)(servers.size() + NumItems - 1) / NumItems);
       args.push_back(msg);
       pageLabel->setString("Page {1} of {2}", &args);
     }
